Add files via upload
This is simple GUI based chat bot.In this one is user and one is Bot one of them can ask one at a time. There is speak button which allow other to speak and one is clear button to clear the chat as this is not storing the previous message.